requirements.txt
Flask
Flask-Cors
google-cloud-speech
google-cloud-storage
google-cloud-aiplatform
vertexai
python-dotenv
numpy
sentence-transformers
einops